requirements.txt
arabic-reshaper==2.0.14
certifi==2018.4.16
chardet==3.0.4
colorama==0.3.9
configparser==3.5.0
dj-database-url==0.4.1
Django==2.0.4
django-webpack-loader==0.6.0
djangorestframework==3.8.2
future==0.16.0
gunicorn==19.6.0
html5lib==1.0.1
httpie==0.9.9
httplib2==0.11.3
idna==2.6
mysqlclient==1.3.12
Pillow==5.1.0
psycopg2==2.7.4
Pygments==2.2.0
PyPDF2==1.26.0
python-bidi==0.4.0
pytz==2018.4
reportlab==3.4.0
requests==2.18.4
six==1.11.0
urllib3==1.22
webencodings==0.5.1
whitenoise==3.2.2
xhtml2pdf==0.2.2
django-rest-knox==3.1.5
djoser==